requirement.txt
fastapi
requests
python-dotenv
apscheduler